Refactor NotificationAPIProvider to accept a client prop and  addNotificationsToState is accessible by the context
This commit refactors the NotificationAPIProvider component to accept a client prop. If the prop is provided, it will be used as the client for the NotificationAPIClientSDK. If the prop is not provided, the component will initialize the client using the existing logic. This change allows for more flexibility in using different client instances with the NotificationAPIProvider.
It also export addNotificationsToState through context